% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{add_integers_R}
\alias{add_integers_R}
\title{Add two integers
'add_integers_R' adds two integers}
\usage{
add_integers_R(a, b)
}
\arguments{
\item{a}{int first integer}

\item{b}{int second integer}
}
\value{
c int sum of a and b
}
\description{
Add two integers
'add_integers_R' adds two integers
}
