% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_last_values.R
\name{add_last_values}
\alias{add_last_values}
\title{Add last values marks to the right of active time series plot}
\usage{
add_last_values(data, ylim, col)
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table} object of plotted data}

\item{ylim}{user specified range of \code{data}}

\item{col}{same as in \code{\link{plot_ts}}}
}
\description{
Add last values marks to the right of active time series plot
}
\details{
Used in \code{\link{plot_ts}} internally.
}
\seealso{
Other graphical functions: \code{\link{add_legend}},
  \code{\link{distinct_colors}}, \code{\link{empty_plot}},
  \code{\link{hist_dt}}, \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_table}},
  \code{\link{plot_ts}}
}
