% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to active time series plot}
\usage{
add_legend(position = "topright", names, col = "auto", lty = 1, lwd = 1,
  pch = NA)
}
\arguments{
\item{position}{same as in \code{\link{plot_ts}} except \code{'n'}}

\item{names}{line labels}

\item{col}{same as in \code{\link{plot_ts}}}

\item{lty, lwd}{same as in \code{\link[graphics]{lines}}}

\item{pch}{same as in \code{\link[graphics]{points}}}
}
\description{
Add legend to active time series plot
}
\details{
Used in \code{\link{plot_ts}} internally.
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{distinct_colors}}, \code{\link{empty_plot}},
  \code{\link{hist_dt}}, \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_table}},
  \code{\link{plot_ts}}
}

