% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{atr}
\alias{atr}
\title{Average True Range}
\usage{
atr(x, n)
}
\arguments{
\item{x}{\code{high, low, close} data.frame}

\item{n}{window size}
}
\value{
data.table with columns \code{tr, atr, trueHigh, trueLow}
}
\description{
Average True Rangeis a measure of volatility of a High-Low-Close series.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{sma}}, \code{\link{stochastic}}
}
\concept{technical indicators}
