% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{back_test}
\alias{back_test}
\title{Generic back test function}
\usage{
back_test(enter, exit, price, stop_loss = -1000, side = 1L)
}
\arguments{
\item{enter}{bool vector of length n of enter signals}

\item{exit}{bool vector of length n of exit signals}

\item{price}{numeric vector of length n of prices}

\item{stop_loss}{relative stop loss, must be negative}

\item{side}{direction of enter order, \code{-1}:short, \code{1}:long}
}
\value{
trades data.table with columns \code{ price_enter,price_exit,mtm_min,mtm_max,id_enter,id_exit,pnl_trade,side}
}
\description{
Back test by enter and exit signals with stop loss on price history. Execution is immediate. Useful for testing on daily data.
}
