% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbands}
\alias{bbands}
\title{Bollinger Bands}
\usage{
bbands(x, n, k)
}
\arguments{
\item{x}{numeric vectors}

\item{n}{window size}

\item{k}{number of standard deviations}
}
\value{
Returns data.table with columns \code{upper, lower, sma}.
}
\description{
Bollinger bands is a mix of Rolling Range and SMA indicators. It shows the average price and its range over n past values based on price volatility.
}
\seealso{
Other technical indicators: \code{\link{crossover}},
  \code{\link{ema}}, \code{\link{roll_lm}},
  \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{sma}}, \code{\link{stochastic}}
}

