% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_decimal_resolution.R
\name{calc_decimal_resolution}
\alias{calc_decimal_resolution}
\title{Calculate decimal resolution}
\usage{
calc_decimal_resolution(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Calculate decimal resolution
}
\details{
Used in \code{\link{add_last_values}} internally.
}
