% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Candle.R
\name{Candle}
\alias{Candle}
\title{C++ Candle class}
\arguments{
\item{id}{id}

\item{open}{price}

\item{high}{price}

\item{low}{price}

\item{close}{price}

\item{time}{seconds since epoch}

\item{volume}{volume}

\item{timeFrame}{timeframe in seconds}
}
\description{
C++ class documentation
}
\section{Usage}{
 \code{Candle{ int id, double open, double high, double low,
double close, double time, int volume, int timeFrame }}
}

\seealso{
Other backtesting classes: \code{\link{Cost}},
  \code{\link{Indicator}}, \code{\link{Order}},
  \code{\link{Processor}}, \code{\link{Tick}}

Other C++ classes: \code{\link{BBands}},
  \code{\link{Cost}}, \code{\link{Crossover}},
  \code{\link{Ema}}, \code{\link{Indicator}},
  \code{\link{Order}}, \code{\link{Processor}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}},
  \code{\link{Tick}}
}
