% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CandleAggregator.R
\name{CandleAggregator}
\alias{CandleAggregator}
\title{C++ Candle Aggregator class}
\arguments{
\item{timeFrame}{candle timeframe}
}
\description{
C++ class documentation
}
\details{
R functions \link{to_candles}.
}
\section{Usage}{
 \code{CandleAggregator( int timeFrame )}
}

\section{Public Members and Methods}{

\tabular{lll}{
\cr \strong{Name}                         \tab \strong{Return Type} \tab \strong{Description}
\cr \code{Add( const \link{Tick}& tick )} \tab \code{void}          \tab update indicator
\cr \code{Reset()}                        \tab \code{void}          \tab reset to initial state
\cr \code{IsFormed()}                     \tab \code{bool}          \tab is indicator value valid?
\cr \code{GetValue()}                     \tab \code{\link{Candle}} \tab see \link{Candle}
\cr \code{GetHistory()}                   \tab \code{List}          \tab return values history data.table with columns \code{time, open, high, low, close, volume, id}
}
}

\seealso{
Other C++ indicators: \code{\link{BBands}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{Indicator}}, \code{\link{RollLinReg}},
  \code{\link{RollPercentRank}}, \code{\link{RollRange}},
  \code{\link{RollSd}}, \code{\link{RollVolumeProfile}},
  \code{\link{Rsi}}, \code{\link{Sma}},
  \code{\link{Stochastic}}

Other C++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Cost}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{Indicator}}, \code{\link{Order}},
  \code{\link{Processor}}, \code{\link{RollLinReg}},
  \code{\link{RollPercentRank}}, \code{\link{RollRange}},
  \code{\link{RollSd}}, \code{\link{RollVolumeProfile}},
  \code{\link{Rsi}}, \code{\link{Sma}},
  \code{\link{Stochastic}}, \code{\link{Tick}}
}
\concept{C++ classes}
\concept{C++ indicators}
