% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Cost.R
\name{Cost}
\alias{Cost}
\title{C++ Trading Commissions class}
\arguments{
\item{pointValue}{price point value ( 1 for stocks )}

\item{cancel}{absolute commission per order cancel}

\item{order}{absolute commission per order}

\item{stockAbs}{absolute commission per stock / contract}

\item{tradeAbs}{absolute commission per trade}

\item{tradeRel}{relative commission per trade volume}

\item{longAbs}{absolute commission/refund per long position}

\item{longRel}{relative commission/refund per long volume}

\item{shortAbs}{absolute commission/refund per short position}

\item{shortRel}{relative commission/refund per short volume}
}
\description{
C++ class documentation
}
\section{Usage}{
 \code{Cost = \{\}}
}

\section{IMPORTANT}{

Positive value means refund, negative value means cost!
}

\seealso{
Other backtesting classes: \code{\link{Candle}},
  \code{\link{Indicator}}, \code{\link{Order}},
  \code{\link{Processor}}, \code{\link{Tick}}

Other C++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Crossover}},
  \code{\link{Ema}}, \code{\link{Indicator}},
  \code{\link{Order}}, \code{\link{Processor}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}},
  \code{\link{Tick}}
}
