% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Crossover.R
\name{Crossover}
\alias{Crossover}
\title{C++ Crossover class}
\description{
C++ class documentation
}
\details{
R function \link{crossover}.
}
\section{Usage}{
 \code{Crossover}
}

\section{Public Members and Methods}{

\tabular{lll}{
 \strong{Name}                                   \tab \strong{Return Type}        \tab \strong{Description}                   \cr
 \code{Add( std::pair< double, double > value )} \tab \code{void}                 \tab update indicator                       \cr
 \code{Reset()}                                  \tab \code{void}                 \tab reset to initial state                 \cr
 \code{IsFormed()}                               \tab \code{bool}                 \tab is indicator value valid?              \cr
 \code{IsAbove()}                                \tab \code{bool}                 \tab first just went above second?          \cr
 \code{IsBelow()}                                \tab \code{bool}                 \tab first just went below second?          \cr
 \code{GetHistory()}                             \tab \code{factor}               \tab factor vector with levels \code{UP, DN}
}
}

\seealso{
Other C++ indicators: \code{\link{BBands}},
  \code{\link{Ema}}, \code{\link{Indicator}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}}

Other C++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Cost}},
  \code{\link{Ema}}, \code{\link{Indicator}},
  \code{\link{Order}}, \code{\link{Processor}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}},
  \code{\link{Tick}}
}
