% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Ema.R
\name{Ema}
\alias{Ema}
\title{C++ Exponential Moving Average class}
\arguments{
\item{n}{indicator period}
}
\description{
C++ class documentation
}
\details{
R function \link{ema}.
}
\section{Usage}{
 \code{Ema( int n )}
}

\section{Public Members and Methods}{

\tabular{lll}{
 \strong{Name}               \tab \strong{Return Type}       \tab \strong{Description}       \cr
 \code{Add( double value )}  \tab \code{void}                \tab update indicator           \cr
 \code{GetValue()}           \tab \code{double}              \tab return value               \cr
 \code{Reset()}              \tab \code{void}                \tab reset to initial state     \cr
 \code{IsFormed()}           \tab \code{bool}                \tab is indicator value valid?  \cr
 \code{GetHistory()}         \tab \code{std::vector<double>} \tab return values history
}
}

\seealso{
Other C++ indicators: \code{\link{BBands}},
  \code{\link{Crossover}}, \code{\link{Indicator}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}}

Other C++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Cost}},
  \code{\link{Crossover}}, \code{\link{Indicator}},
  \code{\link{Order}}, \code{\link{Processor}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}},
  \code{\link{Tick}}
}
