% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Indicator.R
\name{Indicator}
\alias{Indicator}
\title{c++ Indicator Base class}
\arguments{
\item{InputType}{input type}

\item{ValueType}{output type}

\item{HistoryType}{history type}
}
\description{
c++ class documentation
}
\section{Usage}{
 \code{class AnyIndicator : public Indicator< InputType, ValueType, HistoryType > { }}
}

\section{Public Members and Methods}{

Following methods must be specified for \code{AnyIndicator}
\tabular{lll}{
\cr \strong{Name}                           \tab \strong{Return Type} \tab \strong{Description}
\cr \code{Add( InputType input )}           \tab \code{void}          \tab update indicator
\cr \code{Reset()}                          \tab \code{void}          \tab reset to initial state
\cr \code{IsFormed()}                       \tab \code{bool}          \tab is indicator value valid?
\cr \code{GetHistory()}                     \tab \code{HistoryType}   \tab return values history
}
}
\seealso{
Other backtesting classes: \code{\link{Candle}},
  \code{\link{Cost}}, \code{\link{Order}},
  \code{\link{Processor}}, \code{\link{Tick}}

Other c++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Cost}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{Order}}, \code{\link{Processor}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}},
  \code{\link{Tick}}

Other c++ indicators: \code{\link{BBands}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{RollLinReg}}, \code{\link{RollPercentRank}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}}
}

