% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Tick.R
\name{Tick}
\alias{Tick}
\title{C++ Tick class}
\arguments{
\item{id}{id}

\item{time}{seconds since epoch}

\item{price}{price}

\item{volume}{volume}
}
\description{
C++ class documentation
}
\section{Usage}{
 \code{Tick{ int id, double time, double price, int volume }}
}

\seealso{
Other backtesting classes: \code{\link{Candle}},
  \code{\link{Cost}}, \code{\link{Indicator}},
  \code{\link{Order}}, \code{\link{Processor}}

Other C++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Cost}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{Indicator}}, \code{\link{Order}},
  \code{\link{Processor}}, \code{\link{RollLinReg}},
  \code{\link{RollPercentRank}}, \code{\link{RollRange}},
  \code{\link{RollSd}}, \code{\link{RollVolumeProfile}},
  \code{\link{Rsi}}, \code{\link{Sma}},
  \code{\link{Stochastic}}
}
