% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossover}
\alias{crossover}
\title{Crossover}
\usage{
crossover(x, y)
}
\arguments{
\item{x, y}{numeric vectors}
}
\description{
Crossover is binary indicator indicating the moment when one value goes above or below another.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{ema}}, \code{\link{roll_lm}},
  \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{sma}}, \code{\link{stochastic}}
}
