% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum_reset}
\alias{cumsum_reset}
\title{Returns a vector whose elements are the cumulative sums of the inputted
  numeric vector. The argument 'threshold' controls the max/min 
  ceiling/floor of the running sum before it is reset; i.e., if the 
  threshold is set to (+/-) Inf, one would, essentially, compute the base R 
  cumsum', albeit slower}
\usage{
cumsum_reset(x, threshold)
}
\arguments{
\item{x}{A numeric vector}

\item{threshold}{A numeric scalar specifying the cumulative 
threshold(reset)}
}
\description{
Returns a vector whose elements are the cumulative sums of the inputted
  numeric vector. The argument 'threshold' controls the max/min 
  ceiling/floor of the running sum before it is reset; i.e., if the 
  threshold is set to (+/-) Inf, one would, essentially, compute the base R 
  cumsum', albeit slower
}
