% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_colors.R
\docType{data}
\name{distinct_colors}
\alias{distinct_colors}
\title{Distinct colors vector}
\format{An object of class \code{character} of length 25.}
\usage{
distinct_colors
}
\description{
Distinct colors vector
}
\details{
Distinct colors vector.
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{empty_plot}},
  \code{\link{hist_dt}}, \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_table}},
  \code{\link{plot_ts}}
}
\keyword{datasets}
