% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dof.R
\name{dof}
\alias{dof}
\alias{\%dof\%}
\title{Do calculation on data.table excluding first column}
\usage{
dof(obj, action, ...)

obj \%dof\% action
}
\arguments{
\item{obj}{data.table}

\item{action}{function or text string with actions to perform}

\item{...}{additional parameters to function if \code{action} is function}
}
\description{
Do calculation on data.table excluding first column
}
