% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsharpe.R
\name{dsr}
\alias{dsr}
\title{Deflated Sharpe Ratio (DSR)}
\usage{
dsr(N, T, observed.sharpe, variance.sharpe, skew, kurtosis)
}
\arguments{
\item{N}{The total number of trials (parameters) attempted}

\item{T}{The total number of observations (returns) s}

\item{observed.sharpe}{The observed Sharpe Ratio calculated from 'T' real returns}

\item{variance.sharpe}{The observed variance of the Sharpe Ratios's across 'N' trials}

\item{skew}{The observed skew of the 'T' real returns}

\item{kurtosis}{The observed kurtosis of the 'T' real returns}
}
\description{
Computes the deflated sharpe ratio vis-a-vis Extreme Value Theory
}
\details{

}
\examples{
\dontrun{
M               <- replicate(10, rnorm(1000, 0, 1))
trials          <- ncol(M) # number of trials (e.g., parameters attempted)
observations    <- nrow(M) # number of observations (e.g., returns)
evaluation.function <- sharpe.ratio <-
   function(x, rf = 0.02 / 252) mean(x - rf) /  sd(x - rf)

result <- dsr(
     N=observations,
     T=trials,
     observed.sharpe=
)
result

}
}
\references{
Bailey et al (2015) "The Deflated Sharpe Ratio: 
Correcting for Selection Bias, Backtest Overfitting and Non-Normality" 
\url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2460551}
}
\author{
Nathan Matare <email: nmatare@chicagobooth.com>
}
\keyword{dsr}
