% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_plot.R
\name{empty_plot}
\alias{empty_plot}
\title{Plot empty plot}
\usage{
empty_plot()
}
\description{
Plot empty plot
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{distinct_colors}},
  \code{\link{hist_dt}}, \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_table}},
  \code{\link{plot_ts}}
}
