% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_futures_codes.R
\name{gen_futures_codes}
\alias{gen_futures_codes}
\title{Generate futures contract names between dates}
\usage{
gen_futures_codes(contract, from, to, frequency)
}
\arguments{
\item{contract}{contract base name}

\item{from, to}{text dates in format YYYY-MM-DD}

\item{frequency}{expiration frequency, e.g. 3 for quarterly contracts}
}
\description{
Generate futures contract names between dates
}

