% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_dt.R
\name{hist_dt}
\alias{hist_dt}
\title{Plot histogram of data.table by columns}
\usage{
hist_dt(dt, bin_width = diff(range(dt, na.rm = TRUE))/10, coeff = 0.8,
  main = "")
}
\arguments{
\item{dt}{data.table}

\item{bin_width}{truncate data by this value}

\item{coeff}{group width in [0,1]}

\item{main}{plot title}
}
\description{
Plot histogram of data.table by columns
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{distinct_colors}},
  \code{\link{empty_plot}}, \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_table}},
  \code{\link{plot_ts}}
}

