% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investing.R
\name{investing_get_info}
\alias{investing_get_info}
\title{Origin' = 'http://www.investing.com' ), httr::user_agent( paste( 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Ubuntu', 'Chromium/51.0.2704.79 Chrome/51.0.2704.79 Safari/537.36' ) ) )}
\usage{
investing_get_info(url)
}
\description{
Origin' = 'http://www.investing.com' ), httr::user_agent( paste( 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Ubuntu', 'Chromium/51.0.2704.79 Chrome/51.0.2704.79 Safari/537.36' ) ) )
}
