% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iround.R
\name{iround}
\alias{iround}
\title{Round numbers to specified interval}
\usage{
iround(x, interval)
}
\arguments{
\item{x}{numeric vector to be rounded}

\item{interval}{the interval the values should be rounded towards}
}
\value{
A numeric vector with x rounded to the desired interval.
}
\description{
Round numbers to specified interval
}
