% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_named.R
\name{lapply_named}
\alias{lapply_named}
\title{lapply which returns named list}
\usage{
lapply_named(X, FUN, ...)
}
\arguments{
\item{X, FUN, ...}{same as \code{\link[base]{lapply}} arguments}
}
\description{
lapply which returns named list
}
