% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_ohlc.R
\name{lines_ohlc}
\alias{lines_ohlc}
\title{Add candles to active time series plot}
\usage{
lines_ohlc(x = 1:nrow(ohlc), ohlc, width = 0.3, candle.col.up = "blue",
  candle.col.dn = "red", ch = TRUE)
}
\arguments{
\item{x}{location coordinates}

\item{ohlc}{time_series \code{data.frame} or \code{data.table} object with 4 columns \code{'open','high','low','close'}}

\item{width}{width of candles body}

\item{candle.col.up, candle.col.dn}{colors of up and down candles}

\item{ch}{use Chinese style?}
}
\description{
Add candles to active time series plot
}
\details{
Used in \code{\link{plot_ts}} internally.
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{distinct_colors}},
  \code{\link{empty_plot}}, \code{\link{hist_dt}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_table}},
  \code{\link{plot_ts}}
}
