% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_stacked_hist.R
\name{lines_stacked_hist}
\alias{lines_stacked_hist}
\title{Add stacked histogram to active time series plot}
\usage{
lines_stacked_hist(x = 1:nrow(data), data, width = "auto", col = "auto",
  ordered = TRUE)
}
\arguments{
\item{x}{location coordinates}

\item{data}{time_series \code{data.frame} or \code{data.table} object with 4 columns \code{'open','high','low','close'}}

\item{width}{width of histogram segment}

\item{col}{colors of segments}

\item{ordered}{should stacked bars be in order?}
}
\description{
Add stacked histogram to active time series plot
}
\details{
Used in \code{\link{plot_ts}} internally.
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{distinct_colors}},
  \code{\link{empty_plot}}, \code{\link{hist_dt}},
  \code{\link{lines_ohlc}}, \code{\link{multi_heatmap}},
  \code{\link{plot_table}}, \code{\link{plot_ts}}
}

