% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{make_bars}
\alias{make_bars}
\title{'make_bars'}
\usage{
make_bars(x, type, by = c(1, "days"))
}
\arguments{
\item{x}{A data.frame of raw tick data. The object must contain
columns: timestamp(POSIXct), price(numeric), size(numeric), 
and side(character).}

\item{type}{One of either: time, volume, unit, tick, tick runs, tick 
imbalance volume imbalance, unit imbalance, or CUSUM bars. 
See details for further information.}

\item{by}{An integer (or in the case of type='time', an integer 
and interval) specifying the sampling frequency. 
See details for further information.}
}
\description{
Creates financial time series 'bars' (also known as 'candlesticks') 
from raw trade data. One may create either time, volume, unit, tick, 
tick-runs, tick-imbalance (TIBs) volume-imbalance (VIBs), 
unit-imbalance (DIBs), or CUSUM-bars.
}
\details{
'make_bars' creates an OHLCV (Open/High/Low/Close/Volume) matrix of 
class 'xts' The function requires an input of class 'data.frame' or 'matrix' 
with the following columns:

     timestamp:    time of observation; will be used as the xts index 

     price:        the price of the transacted units

     size:         number of transacted units

     side:         an integer vector of 1L(buy) or -1L(sell)

Type may be one of either time, volume, unit, tick runs, tick imbalance
volume imbalance, unit imbalance, or CUSUM bars.

time:       sampled every 'by' units of time; 
            must be a vector of length two, and corresponding to 
            [units, interval];
            for example, 'by=c(1, "days")' or 'by=c(1, "sec")'; 
            default is c(1, "days")

tick:       sampled every 'by' number of ticks

volume:     sampled every 'by' units of volume 

unit:       sampled every 'by' units(dollars) of units (e.g., every $10,000)

*NOTE* 'tick runs', 'tick imbalance', 'volume imbalance', 'unit imbalance', 
and 'CUMSUM' are not yet supported
}
\examples{
\dontrun{
df <- data.frame(
 timestamp=seq(from = as.POSIXct("2018-01-21"), 
               length.out = 100, by = "secs"),
 price=rnorm(100,65,5),
 size=rnorm(100,10,1),
 side=rep(c(1L, -1L), 50)
)

result <- make_bars(x=df, type='time', by=c(5, "mins"))
result

}
}
\references{
Lopez de Prado, Marcos, Advances in Financial Machine 
             Learning (2018)
}
\author{
Nathan Matare <email: nmatare@chicagobooth.com>
}
\keyword{make_bars}
