% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{My R package data}
\format{
An object of class \code{data.frame} with 1000 rows and 1 columns.
}
\usage{
my_data
}
\description{
data created with R base function 'sample'
#' @format ## 'my_data'
A data frame with 1,000 rows and 1 column:
\describe{
\item{col1}{Column 1}
}
}
\keyword{datasets}
