% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_locf.R
\name{na_locf}
\alias{na_locf}
\title{Last Observation Carried Forward}
\usage{
na_locf(x, na = NA)
}
\arguments{
\item{x}{list or vector to roll through}

\item{na}{leading NA substitution}
}
\description{
Last Observation Carried Forward
}
