% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscv.R
\name{plot.cscv}
\alias{plot.cscv}
\title{Graphical analysis of CSCV results
This function supports the plotting of multiple cscv runs viewed in a
aggregate.
cscv.plot() will return:
(1)  the frequency distribution of the rank logits as described in Bailey 
     et al (2015);
(2)  the out-of-sample performance degradation when compared to in-sample
     results given a previously defined evaluation function; 
(3)  the first and second order plots of stochastic dominance present among the
     attempted trials.}
\usage{
\method{plot}{cscv}(cscv.objects, summarize = 50)
}
\arguments{
\item{cscv.objects}{A list containing the results of a particular cscv run. 
This plotting supports multiple cscv runs so one may 
view the results in aggregate.}

\item{summarize}{Whether to summarize the R/Rbar pairs. Helpful when 
S >> 16 and one has many pairs across many runs; 
default 50.}
}
\description{
The legend includes the attributed 'phi' and 'beta' which
correspond to the probability of overfit and the slope of OOS performance 
degradation line, respectively.
}
\seealso{
cscv
}
\keyword{cscv}
