% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot data.frame as table with histogram in background}
\usage{
plot_table(x, hist = "n", col = "auto", srt = 0, ...)
}
\arguments{
\item{x}{data}

\item{hist}{histogram background type, \code{'bycol', 'total', 'n'}}

\item{col}{only auto colors available}

\item{srt}{column names rotation}

\item{...}{further graphical parameters as in \code{\link[graphics]{par}}}
}
\description{
Plot data.frame as table with histogram in background
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{distinct_colors}},
  \code{\link{empty_plot}}, \code{\link{hist_dt}},
  \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_ts}}
}

