% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_market_regime.R
\name{qt_market_regime}
\alias{qt_market_regime}
\title{qt_market_regime}
\usage{
qt_market_regime(x, days = 100)
}
\arguments{
\item{x}{A numeric vector of daily returns}

\item{days}{Period of days to calculate}
}
\value{
A string of the current market regime
}
\description{
qt_market_regime
}
