% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_percentile.R
\name{qt_percentile}
\alias{qt_percentile}
\title{qt_percentile}
\usage{
qt_percentile(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric value
}
\description{
Calculates the percentile of the last value isnide a numeric vector
}
