% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_stochastic_oscillator.R
\name{qt_stochastic_oscillator}
\alias{qt_stochastic_oscillator}
\title{qt_stochastic_oscillator}
\usage{
qt_stochastic_oscillator(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric value
}
\description{
Returns the stochastic oscillator calculation of the last value of a numeric vector
}
