% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_zscore.R
\name{qt_zscore}
\alias{qt_zscore}
\title{qt_zscore}
\usage{
qt_zscore(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric value
}
\description{
Calculates the Z-Score of the last value inside a numeric vector
}
