% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantTools-package.R
\docType{package}
\name{QuantTools-package}
\alias{QuantTools}
\alias{QuantTools-package}
\title{QuantTools: An R Package for Estimating and Testing Asset Pricing Factor Models}
\description{
The proposed package is designed to provide a comprehensive toolset for financial econometrics, focusing on the estimation and testing of asset pricing factor models. By incorporating a variety of typical asset pricing datasets and efficient implementation of state-of-the-art methodologies, this package aims to simplify the process of data preparation and methodology selection. The package introduces features for the estimation of factor models and testing for model misspecification, model identification, and the significance of risk premia. Moreover, it incorporates various regression techniques, Continuous Update Estimation (CUE), Generalized Method of Moments (GMM) techniques, and a penalized approach.
}
\author{
\strong{Maintainer}: Ali Moin \email{a.moin@tinbergen.nl} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
