% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{returns}
\alias{returns}
\title{Calculate returns}
\usage{
returns(x, type = "r", n = 1)
}
\arguments{
\item{x}{numeric vector}

\item{type}{\code{'r' = x[t] / x[t-n] - 1 }, \code{'l' = ln( x[t] / x[t-n] ) }}

\item{n}{lookback}
}
\value{
Vector of same length as x with absent returns converted to 0 for relative and 1 for logarithmic.
}
\description{
Calculate returns
}
