% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_futures.R
\name{roll_futures}
\alias{roll_futures}
\title{Combine multiple futures market data into continuous contract}
\usage{
roll_futures(prices_by_contract, days_before_expiry)
}
\arguments{
\item{prices_by_contract}{list of data.tables with futures market data}

\item{days_before_expiry}{number of dates before expiration to roll}
}
\description{
Combine multiple futures market data into continuous contract
}
