% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_lm}
\alias{roll_lm}
\alias{roll_correlation}
\title{Rolling Linear Regression}
\usage{
roll_lm(x, y, n)

roll_correlation(x, y, n)
}
\arguments{
\item{x, y}{numeric vectors}

\item{n}{window size}
}
\value{
roll_lm returns data.table with columns \code{alpha, beta, r, r.squared}
}
\description{
Rolling linear regression calculates regression coefficients over n past paired values.
\cr Others return numeric vector
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{sma}}, \code{\link{stochastic}}
}
