% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_percent_rank}
\alias{roll_percent_rank}
\title{Rolling Percent Rank}
\usage{
roll_percent_rank(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{window size}
}
\description{
Rolling percent rank normalizes values to a range from 0 to 100.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_range}},
  \code{\link{roll_sd}}, \code{\link{roll_volume_profile}},
  \code{\link{rsi}}, \code{\link{sma}},
  \code{\link{stochastic}}
}

