% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_range}
\alias{roll_max}
\alias{roll_min}
\alias{roll_quantile}
\alias{roll_range}
\title{Rolling Range}
\usage{
roll_range(x, n)

roll_quantile(x, n, p)

roll_min(x, n)

roll_max(x, n)
}
\arguments{
\item{x}{numeric vectors}

\item{n}{window size}

\item{p}{probability value \code{[0, 1]}}
}
\value{
roll_range returns data.table with columns \code{min, max}
\cr others return numeric vector
}
\description{
Rolling range is minimum and maximum values over n past values. Can be used to identify price range.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_percent_rank}},
  \code{\link{roll_sd}}, \code{\link{roll_volume_profile}},
  \code{\link{rsi}}, \code{\link{sma}},
  \code{\link{stochastic}}
}

