% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sd}
\alias{roll_sd}
\title{Rolling Standard Deviation}
\usage{
roll_sd(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{window size}
}
\description{
Rolling standard deviation shows standard deviation over n past values.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_percent_rank}},
  \code{\link{roll_range}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{sma}}, \code{\link{stochastic}}
}

