% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sd_filter}
\alias{roll_sd_filter}
\title{Rolling Filter}
\usage{
roll_sd_filter(x, n, k = 1, m = 10L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{window size}

\item{k}{number of standard deviations}

\item{m}{number of consequent large returns to stop filtering out}
}
\description{
Logical vector is returned. This function is useful to filter ticks. Finds consequent elements which absolute change is higher than k standard deviation of past n changes and mark them \code{FALSE}. If sequence length greater than \code{m} values become \code{TRUE}.
}
