% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_POSIXct.R
\name{round_POSIXct}
\alias{round_POSIXct}
\title{Round POSIXct timestamps}
\usage{
round_POSIXct(x, units = c("seconds", "minutes", "5min", "10min", "15min",
  "30min", "hour", "day"), type = "floor")
}
\arguments{
\item{x}{POSIXct vector}

\item{units}{to round off to}

\item{type}{\code{floor} or \code{ceiling}}
}
\description{
Round POSIXct timestamps
}
\details{
Rounds POSIXct vector.
}

