% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsi}
\alias{rsi}
\title{Relative Strength Index}
\usage{
rsi(x, n)
}
\arguments{
\item{x}{numeric vectors}

\item{n}{window size}
}
\description{
Relative strength index aka RSI measures the velocity and magnitude of directional price movements.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{sma}},
  \code{\link{stochastic}}
}
