% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sma}
\alias{sma}
\title{Simple Moving Average}
\usage{
sma(x, n)
}
\arguments{
\item{x}{numeric vectors}

\item{n}{window size}
}
\description{
Simple moving average also called SMA is the most popular indicator. It shows the average of n past values. Can be used for time series smoothing.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{stochastic}}
}
