% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stochastic}
\alias{stochastic}
\title{Stochastic}
\usage{
stochastic(x, n, nFast, nSlow)
}
\arguments{
\item{x}{\code{high, low, close} data.frame or numeric vector}

\item{n}{window size}

\item{nFast}{fast smooth}

\item{nSlow}{slow smooth}
}
\value{
data.table with columns \code{k_fast, d_fast, d_slow}
}
\description{
Stochastic oscillator shows position of price in respect to its range over n past values.
}
\seealso{
Other technical indicators: \code{\link{bbands}},
  \code{\link{crossover}}, \code{\link{ema}},
  \code{\link{roll_lm}}, \code{\link{roll_percent_rank}},
  \code{\link{roll_range}}, \code{\link{roll_sd}},
  \code{\link{roll_volume_profile}}, \code{\link{rsi}},
  \code{\link{sma}}
}

