% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlt.R
\name{tlt}
\alias{tlt}
\title{Lapply function to data.table by row.}
\usage{
tlt(x, fun, ...)
}
\arguments{
\item{x}{data.table}

\item{fun}{function}
}
\description{
T.ranspose L.apply T.ranspose algorithm used. First step is to transpose initial data.table so rows become columns. Second is to lapply function to new columns. And the last step is to transpose back to original shape.
}
