% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_candles}
\alias{to_candles}
\title{Convert ticks to candles}
\usage{
to_candles(ticks, timeframe)
}
\arguments{
\item{ticks}{read 'Ticks' section in \link{Processor}}

\item{timeframe}{candle timeframe in seconds}
}
\value{
data.table with columns \code{time, open, high, low, close, volume, id}. Where \code{id} is row number of last tick in candle. \cr
Note: last candle is always omitted.
}
\description{
Convert ticks to candles
}
