% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_UTC.R
\name{to_UTC}
\alias{to_UTC}
\title{Convert time zone to 'UTC' without changing value}
\usage{
to_UTC(x)
}
\arguments{
\item{x}{\code{POSIXct} vector}
}
\description{
Convert time zone to 'UTC' without changing value
}
\examples{
\donttest{

Sys.time()
to_UTC( Sys.time() )

}

}
