% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vswitch.R
\name{vswitch}
\alias{vswitch}
\title{Vectorized switch}
\usage{
vswitch(x, mapping, default = NA)
}
\arguments{
\item{x}{old values}

\item{mapping}{named vector mapping \code{c( old_names = new_names )}}

\item{default}{default value if no mapping found, if set to NULL original value kept}
}
\description{
Vectorized switch
}
