% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wo.R
\name{wo}
\alias{wo}
\alias{\%w/o\%}
\title{Select values in one vector not present in another}
\usage{
x \%w/o\% y
}
\arguments{
\item{x, y}{vectors}
}
\value{
x elements without y elements
}
\description{
Select values in one vector not present in another
}
