% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_RollPercentRank.R
\name{RollPercentRank}
\alias{RollPercentRank}
\title{c++ Rolling Percent Rank class}
\arguments{
\item{n}{indicator period}
}
\description{
c++ class documentation
}
\section{Usage}{
 \code{RollPercentRank( int n )}
}

\section{Public Members and Methods}{

\tabular{lll}{
\cr \strong{Name}                 \tab \strong{Return Type}        \tab \strong{Description}
\cr \code{Add( InputType value )} \tab \code{void}                 \tab update indicator
\cr \code{Reset()}                \tab \code{void}                 \tab reset to initial state
\cr \code{IsFormed()}             \tab \code{bool}                 \tab is indicator value valid?
\cr \code{GetHistory()}           \tab \code{std::vector<double>}  \tab history vector
}
}
\seealso{
Other c++ classes: \code{\link{BBands}},
  \code{\link{Candle}}, \code{\link{Cost}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{Indicator}}, \code{\link{Order}},
  \code{\link{Processor}}, \code{\link{RollLinReg}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}},
  \code{\link{Tick}}

Other c++ indicators: \code{\link{BBands}},
  \code{\link{Crossover}}, \code{\link{Ema}},
  \code{\link{Indicator}}, \code{\link{RollLinReg}},
  \code{\link{RollRange}}, \code{\link{RollSd}},
  \code{\link{RollVolumeProfile}}, \code{\link{Rsi}},
  \code{\link{Sma}}, \code{\link{Stochastic}}
}

