% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot data.table as table}
\usage{
plot_table(dt, transpose = F, justify = c("middle", "left", "right"), ...)
}
\arguments{
\item{dt}{data.table}

\item{transpose}{should table be transposed?}

\item{justify}{\code{'middle','left','right'}}

\item{...}{further graphical parameters as in \code{\link[graphics]{par}}}
}
\description{
Plot data.table as table
}
\seealso{
Other graphical functions: \code{\link{add_last_values}},
  \code{\link{add_legend}}, \code{\link{distinct_colors}},
  \code{\link{empty_plot}}, \code{\link{hist_dt}},
  \code{\link{lines_ohlc}},
  \code{\link{lines_stacked_hist}},
  \code{\link{multi_heatmap}}, \code{\link{plot_ts}}
}
