% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_beta.R
\name{qt_beta}
\alias{qt_beta}
\title{qt_beta}
\usage{
qt_beta(asset_1, asset_2)
}
\arguments{
\item{asset_1}{Independent Variable}

\item{asset_2}{Dependent Variable}
}
\value{
A numeric value: The Beta
}
\description{
Calculates the beta of two price time series.
}
