% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_POSIXct.R
\name{round_POSIXct}
\alias{round_POSIXct}
\alias{ceiling_POSIXct}
\alias{trunc_POSIXct}
\title{Round POSIXct timestamps}
\usage{
round_POSIXct(x, n = 1, units = c("secs", "mins", "hours", "days"),
  method = round)

ceiling_POSIXct(x, n = 1, units = c("secs", "mins", "hours", "days"))

trunc_POSIXct(x, n = 1, units = c("secs", "mins", "hours", "days"))
}
\arguments{
\item{x}{POSIXct vector}

\item{n}{number of units to round off}

\item{units}{to round off to}

\item{method}{round method, see \link[base]{Round}}
}
\description{
Round POSIXct timestamps
}
\details{
Rounds POSIXct vector with specified method.
}
